Integrate ArduCam module and database
