+++ pi-aircraft-detector.py
@@ -29,13 +29,15 @@ import numpy as np
 import time
 import os
 import datetime
-import sqlite3
 import threading
 import argparse
 import logging
 from flask import Flask, Response, render_template
 import json
 import math
+
+from arducam_camera import ArduCam64MP as Camera
+from database import Database
 
 # Configure logging
 logging.basicConfig(
@@ -159,202 +161,7 @@ class AircraftTracker:
                 
         return self.objects
 
-class Camera:
-    """Interface for the 64MP ArduCam with autofocus"""
-    
-    def __init__(self, resolution=(4624, 3472), framerate=15):
-        """
-        Initialize the ArduCam camera
-        
-        Note: The actual resolution used may be lower depending on processing power.
-        For real-time processing, you might want to use a lower resolution.
-        """
-        self.resolution = resolution
-        self.framerate = framerate
-        self.camera = None
-        
-    def initialize(self):
-        """Initialize the camera hardware"""
-        try:
-            # For ArduCam, you need to use their specific library
-            # Below is a placeholder - replace with actual ArduCam initialization
-            
-            # For development/testing, we'll use OpenCV's VideoCapture
-            self.camera = cv2.VideoCapture(0)  # Use 0 for default camera
-            
-            # Set resolution
-            self.camera.set(cv2.CAP_PROP_FRAME_WIDTH, self.resolution[0])
-            self.camera.set(cv2.CAP_PROP_FRAME_HEIGHT, self.resolution[1])
-            
-            # Set framerate
-            self.camera.set(cv2.CAP_PROP_FPS, self.framerate)
-            
-            # Allow camera to initialize
-            time.sleep(2)
-            
-            logger.info(f"Camera initialized at resolution {self.resolution}")
-            return True
-        except Exception as e:
-            logger.error(f"Failed to initialize camera: {e}")
-            return False
-            
-    def release(self):
-        """Release the camera resources"""
-        if self.camera:
-            self.camera.release()
-            logger.info("Camera resources released")
-            
-    def capture_frame(self):
-        """Capture a frame from the camera"""
-        if not self.camera:
-            logger.error("Camera not initialized")
-            return None
-            
-        ret, frame = self.camera.read()
-        if not ret:
-            logger.warning("Failed to capture frame")
-            return None
-            
-        return frame
-        
-    def set_autofocus(self, enable=True):
-        """Enable or disable autofocus"""
-        # This would need to be implemented using ArduCam's specific API
-        # Below is a placeholder
-        if enable:
-            logger.info("Autofocus enabled")
-        else:
-            logger.info("Autofocus disabled")
-            
-    def adjust_exposure(self, sky_brightness):
-        """Adjust camera exposure based on sky brightness"""
-        # This would need to be implemented using ArduCam's specific API
-        # Below is a placeholder
-        logger.info(f"Adjusted exposure for sky brightness: {sky_brightness}")
 
-class Database:
-    """SQLite database for storing aircraft detections"""
-    
-    def __init__(self, db_path):
-        self.db_path = db_path
-        self.conn = None
-        
-    def initialize(self):
-        """Initialize the database and create tables if they don't exist"""
-        try:
-            self.conn = sqlite3.connect(self.db_path, check_same_thread=False)
-            cursor = self.conn.cursor()
-            
-            # Create detections table
-            cursor.execute('''
-                CREATE TABLE IF NOT EXISTS detections (
-                    id INTEGER PRIMARY KEY AUTOINCREMENT,
-                    timestamp TEXT NOT NULL,
-                    x INTEGER NOT NULL,
-                    y INTEGER NOT NULL,
-                    width INTEGER NOT NULL,
-                    height INTEGER NOT NULL,
-                    contrast REAL NOT NULL,
-                    confidence REAL NOT NULL,
-                    image_path TEXT,
-                    speed REAL,
-                    direction REAL
-                )
-            ''')
-            
-            # Create tracking table
-            cursor.execute('''
-                CREATE TABLE IF NOT EXISTS tracking (
-                    id INTEGER PRIMARY KEY AUTOINCREMENT,
-                    detection_id INTEGER NOT NULL,
-                    timestamp TEXT NOT NULL,
-                    x INTEGER NOT NULL,
-                    y INTEGER NOT NULL,
-                    FOREIGN KEY (detection_id) REFERENCES detections (id)
-                )
-            ''')
-            
-            self.conn.commit()
-            logger.info("Database initialized successfully")
-            return True
-        except Exception as e:
-            logger.error(f"Database initialization failed: {e}")
-            return False
-            
-    def close(self):
-        """Close the database connection"""
-        if self.conn:
-            self.conn.close()
-            logger.info("Database connection closed")
-            
-    def record_detection(self, x, y, width, height, contrast, confidence, image_path=None, speed=None, direction=None):
-        """Record a new aircraft detection"""
-        if not self.conn:
-            logger.error("Database not initialized")
-            return None
-            
-        try:
-            cursor = self.conn.cursor()
-            timestamp = datetime.datetime.now().isoformat()
-            
-            cursor.execute('''
-                INSERT INTO detections 
-                (timestamp, x, y, width, height, contrast, confidence, image_path, speed, direction)
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            ''', (timestamp, x, y, width, height, contrast, confidence, image_path, speed, direction))
-            
-            self.conn.commit()
-            return cursor.lastrowid
-        except Exception as e:
-            logger.error(f"Failed to record detection: {e}")
-            return None
-            
-    def record_tracking(self, detection_id, x, y):
-        """Record a tracking update for an existing detection"""
-        if not self.conn:
-            logger.error("Database not initialized")
-            return False
-            
-        try:
-            cursor = self.conn.cursor()
-            timestamp = datetime.datetime.now().isoformat()
-            
-            cursor.execute('''
-                INSERT INTO tracking
-                (detection_id, timestamp, x, y)
-                VALUES (?, ?, ?, ?)
-            ''', (detection_id, timestamp, x, y))
-            
-            self.conn.commit()
-            return True
-        except Exception as e:
-            logger.error(f"Failed to record tracking: {e}")
-            return False
-            
-    def get_recent_detections(self, limit=100):
-        """Get recent aircraft detections"""
-        if not self.conn:
-            logger.error("Database not initialized")
-            return []
-            
-        try:
-            cursor = self.conn.cursor()
-            cursor.execute('''
-                SELECT * FROM detections
-                ORDER BY timestamp DESC
-                LIMIT ?
-            ''', (limit,))
-            
-            columns = [column[0] for column in cursor.description]
-            detections = []
-            
-            for row in cursor.fetchall():
-                detections.append(dict(zip(columns, row)))
-                
-            return detections
-        except Exception as e:
-            logger.error(f"Failed to get recent detections: {e}")
-            return []
 
 class ImageProcessor:
     """Processes camera frames to detect aircraft"""
